feature: Code improved
