+++ app/src/main/java/com/jh/holaflytest/ui/navigation/HolaflyTestNavigation.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.jh.holaflytest.ui.screens.SplashScreen
+import com.jh.holaflytest.ui.screens.splash.SplashScreen
 
 /**
  * Created by Jorge Henao on 13/11/23.
