--- app/src/main/java/com/jh/holaflytest/ui/screens/SplashScreen.kt
@@ -1,86 +1,0 @@
-package com.jh.holaflytest.ui.screens
-
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.basicMarquee
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.ElevatedButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.jh.holaflytest.R
-import com.jh.holaflytest.ui.navigation.Screens
-import com.jh.holaflytest.ui.screens.components.TitleText
-
-/**
- * Created by Jorge Henao on 14/11/23.
- */
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun SplashScreen(
-    onNavigateScreen: (String) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(
-                Brush.verticalGradient(
-                    colors = listOf(
-                        MaterialTheme.colorScheme.primary,
-                        MaterialTheme.colorScheme.onSecondary,
-                    )
-                )
-            ),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.SpaceBetween
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.marvel_avengers),
-            contentDescription = "splash_image",
-            contentScale = ContentScale.Crop,
-            modifier = Modifier
-                .padding(top = 100.dp)
-                .size(300.dp)
-                .clip(RoundedCornerShape(16.dp))
-        )
-        ElevatedButton(
-            modifier = Modifier
-                .padding(24.dp)
-                .fillMaxWidth(),
-            colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.onSurface),
-            onClick = { onNavigateScreen(Screens.HomeScreen.name) }
-        ) {
-            Text(
-                stringResource(id = R.string.splash_start),
-                color = MaterialTheme.colorScheme.onBackground
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun SplashScreenPreview() {
-    SplashScreen {
-
-    }
-}
