+++ app/src/main/java/com/jh/holaflytest/ui/screens/splash/SplashScreen.kt
@@ -1,0 +1,48 @@
+package com.jh.holaflytest.ui.screens.splash
+
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.tooling.preview.Preview
+import com.jh.holaflytest.ui.navigation.Screens
+
+/**
+ * Created by Jorge Henao on 14/11/23.
+ */
+@Composable
+fun SplashScreen(
+    onNavigateScreen: (String) -> Unit
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(
+                Brush.verticalGradient(
+                    colors = listOf(
+                        MaterialTheme.colorScheme.primary,
+                        MaterialTheme.colorScheme.onSecondary,
+                    )
+                )
+            ),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.SpaceBetween
+    ) {
+        SplashHeader()
+        SplashAction {
+            onNavigateScreen(Screens.HomeScreen.name)
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun SplashScreenPreview() {
+    SplashScreen {}
+}
